Move temporal localization from the reasoner into the segment classifier
Cosmos3EdgeReasoner is a pure VLM again (prompt / prompt_video only).
Cosmos3VideoSegmentClassification owns the temporal-localization prompt
template and the strict parser, which now returns
VideoSegmentClassification dataclasses directly, and calls the
reasoner's generic prompt_video. The Phase 0 integration gate targets
the segment classifier's infer().
